Refactor variant and number literal serialization
The commit moves the methods for variant and number literal serializations from the ResourceSerializer class to the VariantSerializer class. The move aids in organization and code readability, as variant serialization is now better isolated and number literal serialization is directly utilized in variant serialization. Exception messages for invalid inputs have also been improved during the refactoing.